Refactor WrappingHBox and busy_check to own modules
